Input validation for Car Info
- ensures that the user inputs are checked for however they are set to be limited before calling database to ensure no malicious calls could be made.
- Remove datacontext as the sql queries uses DatabaseService instead.

+++ ProjectCarTest/ProjectCarTest/Controllers/CarInfoController.cs
@@ -5,6 +5,7 @@ using ProjectCarTest.Dto;
 using System.Security.Claims;
 using System.Collections.Generic;
 using System.Linq;
+using static ProjectCarTest.Utilities.Helper;
 
 // Handles HTTP requests and responses for Car Information for a particular logged-in user via GET, POST, PUT, DELETE requests
 namespace CarTest.Controllers
@@ -14,6 +15,11 @@ namespace CarTest.Controllers
     public class CarInfoController : ControllerBase
     {
         private readonly ICarInfoRepository _carInfoRepository;
+        // Default Parameter Limits
+        private int parameterMaxLength = 100;
+        private int maxStockCount = 999999;
+        private int minYear = 1855;
+        private int currentYear = DateTime.Now.Year;
 
         public CarInfoController(ICarInfoRepository carInfoRepository)
         {
@@ -34,6 +40,9 @@ namespace CarTest.Controllers
         public IActionResult GetCarsByUser()
         {
             var userID = GetUserIdFromToken();
+            if (userID == -1)
+                return Unauthorized("Token is missing or invalid.");
+
             var cars = _carInfoRepository.GetCarsByUserId(userID);
             if (cars == null || !cars.Any()) return NotFound();
             return Ok(cars);
@@ -42,19 +51,35 @@ namespace CarTest.Controllers
         [HttpGet("make/{make}")]
         public IActionResult GetCarsByMake(string make)
         {
-            // TODO: check the following
-            // Make is within 100 char limit, english words, doesn't contain restricted characters, not null
-            var userID = GetUserIdFromToken();
-            var cars = _carInfoRepository.GetCarsByMake(userID, make);
+            // Input Validation
+            if (make.Length > parameterMaxLength) // not accept if input exceeds defined max size
+                return BadRequest("Invalid input for make - Input Length exceeded String Length!");
+            if (!StringValidator.ContainsOnlyLegalCharacters(make)) // not accept any illegal characters
+                return BadRequest("Invalid input for make - Input contains Illegal Characters!");
+
+            var userId = GetUserIdFromToken();
+            if (userId == -1)
+                return Unauthorized("Token is missing or invalid.");
+
+            var cars = _carInfoRepository.GetCarsByMake(userId, make);
+
             return Ok(cars);
         }
 
         [HttpGet("model/{model}")]
         public IActionResult GetCarsByModel(string model)
         {
-            // TODO: check the following
-            // Model is within 100 char limit, english words, doesn't contain restricted characters, not null
+            // Input Validation
+            if (model.Length > parameterMaxLength) // not accept if input exceeds defined max size
+                return BadRequest("Invalid input for model - Input Length exceeded String Length!");
+            if (!StringValidator.ContainsOnlyLegalCharacters(model)) // not accept any illegal characters
+                return BadRequest("Invalid input for model - Input contains Illegal Characters!");
+
             var userID = GetUserIdFromToken();
+            if (userID == -1)
+                return Unauthorized("Token is missing or invalid.");
+
+
             var cars = _carInfoRepository.GetCarsByModel(userID, model);
             return Ok(cars);
         }
@@ -62,10 +87,20 @@ namespace CarTest.Controllers
         [HttpGet("make/{make}/model/{model}")]
         public IActionResult GetCarsByMakeAndModel(string make, string model)
         {
-            // TODO: check the following
-            // Make is within 100 char limit, english words, doesn't contain restricted characters, not null
-            // Model is within 100 char limit, english words, doesn't contain restricted characters, not null
+            // Input Validation
+            if (make.Length > parameterMaxLength) // not accept if make exceeds defined max size
+                return BadRequest("Invalid input for make - Input Length exceeded String Length!");
+            if (!StringValidator.ContainsOnlyLegalCharacters(make)) // not accept any illegal characters
+                return BadRequest("Invalid input for make - Input contains Illegal Characters!");
+            if (model.Length > parameterMaxLength) // not accept if model exceeds defined max size
+                return BadRequest("Invalid input for model - Input Length exceeded String Length!");
+            if (!StringValidator.ContainsOnlyLegalCharacters(model)) // not accept any illegal characters
+                return BadRequest("Invalid input for model - Input contains Illegal Characters!");
+
             var userID = GetUserIdFromToken();
+            if (userID == -1)
+                return Unauthorized("Token is missing or invalid.");
+
             var cars = _carInfoRepository.GetCarsByMakeAndModel(userID, make, model);
             return Ok(cars);
         }
@@ -74,13 +109,27 @@ namespace CarTest.Controllers
         [HttpPost("add")]
         public IActionResult AddCar([FromBody] CarCreateDto carDto)
         {
-            // TODO: check if carId is within bounds
+            // ## Input Validation
             // Make is within 100 char limit, english words, doesn't contain restricted characters, not null
             // Model is within 100 char limit, english words, doesn't contain restricted characters, not null
             // Year is between 1855-<present> (first car made in 1855, probably will never have a car with that date but never know), not null
             // Stock level is non-negative and between 1-<stock limit>
+            if (carDto.Make.Length > parameterMaxLength) // not accept if make exceeds defined max size
+                return BadRequest("Invalid input for make - Input Length exceeded String Length!");
+            if (!StringValidator.ContainsOnlyLegalCharacters(carDto.Make)) // not accept any illegal characters
+                return BadRequest("Invalid input for make - Input contains Illegal Characters!");
+            if (carDto.Model.Length > parameterMaxLength) // not accept if model exceeds defined max size
+                return BadRequest("Invalid input for model - Input Length exceeded String Length!");
+            if (!StringValidator.ContainsOnlyLegalCharacters(carDto.Model)) // not accept any illegal characters
+                return BadRequest("Invalid input for model - Input contains Illegal Characters!");
+            if (carDto.Year < minYear || carDto.Year > currentYear)
+                return BadRequest($"Invalid input for year - Input must be between {minYear}-{currentYear}");
+            if (carDto.StockLevel < 0 || carDto.StockLevel > maxStockCount) // must be between 0-<max stock count>
+                return BadRequest($"Invalid input for stock level - Input is not between 0 and {maxStockCount}!");
 
             var userID = GetUserIdFromToken();
+            if (userID == -1)
+                return Unauthorized("Token is missing or invalid.");
 
             var car = new CarInfo
             {
@@ -109,9 +158,10 @@ namespace CarTest.Controllers
         [HttpDelete("remove/{carId}")]
         public IActionResult RemoveCar(int carId)
         {
-            // TODO: check if carId is within bounds (not negative, is number, not null)
-
             var userID = GetUserIdFromToken();
+            if (userID == -1)
+                return Unauthorized("Token is missing or invalid.");
+
             var success = _carInfoRepository.RemoveCar(carId, userID);
             if (!success) // fails if the carID isn't owned by the user or isn't present in the database.
                 return NotFound($"Car with ID {carId} for the logged-in user not found.");
@@ -122,9 +172,13 @@ namespace CarTest.Controllers
         [HttpPut("update-stock/{carId}")]
         public IActionResult UpdateStockLevel(int carId, [FromBody] CarInfoUpdateStockDto dto)
         {
-            // TODO: check if StockLevel is within bounds (not negative, is number, not null)
+            if (dto.StockLevel < 0 || dto.StockLevel > maxStockCount)
+                return BadRequest($"Invalid input for stock level - Input is not between 0 and {maxStockCount}!");
 
             var userID = GetUserIdFromToken();
+            if (userID == -1)
+                return Unauthorized("Token is missing or invalid.");
+
             var success = _carInfoRepository.UpdateStockLevel(carId, userID, dto.StockLevel);
 
             if (!success) // could not update the stock level
@@ -144,7 +198,12 @@ namespace CarTest.Controllers
         private int GetUserIdFromToken()
         {
             var userIdString = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
-            return int.Parse(userIdString);
+            if (string.IsNullOrEmpty(userIdString) || !int.TryParse(userIdString, out int userId))
+                return -1; // -1 is invalid userID
+
+            return userId;
         }
+
+
     }
 }
